Initial Commit
Creating the skeleton of a card battler. This initial commit contains a mostly empty campaign with a skeleton of a game loop. It is currently untested as it does not currently function as desired, and will need some tweaking to function. Creating a commit now to utilize source control as I play around with the files.